% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{catdrought_app}
\alias{catdrought_app}
\title{function to launch the catdrought app}
\usage{
catdrought_app(user = "ifn", password = "IFN2018creaf", host = NULL,
  port = NULL, dbname = "catdrought_db")
}
\description{
function to launch the catdrought app
}
