% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_update.R
\name{catdrought_daily_update}
\alias{catdrought_daily_update}
\title{Function to update the database with the dates missing}
\usage{
catdrought_daily_update(user = "guest", pass = "guest",
  port = "staging",
  path = "/home/vgranda/LFC/10_serverproces_data/CatDrought",
  changed = TRUE, added = TRUE)
}
\arguments{
\item{added}{}
}
\description{
Usually, this function will update the database with the daily data from serverprocess
}
